// src/app/Models/IS_UOS_SIGNUP.ts
export interface IS_UOS_SIGNUP {
    USU_ID: number;
    USU_ACCCODE: number | null;
    USU_ACCOUNTTYPE: string;
    USU_TITLE: string;
    USU_CNIC: string;
    USU_EMAILADDRESS: string;
    USU_CELLNUMBER: string;
    USU_USERID: string;
    USU_DOB: string | null;
    USU_TRACKID: string;
    USU_SECRETQUESTION: string;
    USU_SECRETANSWER: string;
    USU_STATUS: string;
    UOS_CREATEDON: string | null;
    USU_COMMENTS: string;
    USU_NTNNO: string;
    USU_AUTHORIZEDPERSON: string;
    USU_INCORPORATIONNO: string;
    USU_AUTHORIZEDPERSONDOB: string | null;
    USU_AUTHORIZEDPERSONCNIC: string;
    USU_ADDRESSLINE1: string;
}